Update VID/PIDs from xpad driver in Linux 3.19-rc2
xboxdrv was missing the VID/PID for my Razer Sabertooth. I noticed that
xpad kernel driver recently added the VID/PID for this version of the
controller so I copied that definition. In the same go I also copied
definitions of other controllers that were present in the kernel driver
but missing from xboxdrv.